# openAPI.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Dataloaders API",
        "description": "API documentation for Dataloaders application",
        "contact": {
            "name": "Dataloaders Team",
            "email": "support@dataloaders.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8083",
            "description": "Generated server url"
        }
    ],
...
    "paths": {
...
        "/api/v1/dataloader/configFields/reorder": {
            "post": {
                "tags": [
                    "Field Definitions"
                ],
                "summary": "Bulk reorder field definitions",
                "description": "Reorder multiple field definitions",
                "operationId": "bulkReorder",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/FieldDefinition"
                                }
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Field definitions reordered successfully",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Response"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/dataloader/resource/{resourceId}/status": {
            "patch": {
                "tags": [
                    "Resources"
                ],
                "summary": "Update resource status",
                "description": "Activate or deactivate a resource",
                "operationId": "updateStatus",
                "parameters": [
                    {
                        "name": "resourceId",
                        "in": "path",
                        "description": "Resource ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "isActive",
                        "in": "query",
                        "description": "Active status",
                        "required": true,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Status updated successfully",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Response"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/dataloader/job/{jobId}/status": {
            "patch": {
                "tags": [
                    "Job Config"
                ],
                "summary": "Update job config status",
                "description": "Update job configuration status",
                "operationId": "updateStatus_2",
                "parameters": [
                    {
                        "name": "jobId",
                        "in": "path",
                        "description": "Job ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "isActive",
                        "in": "query",
                        "description": "Active status",
                        "required": true,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Status updated successfully",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Response"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/dataloader/configFields/{resourceTypeId}/status": {
            "patch": {
                "tags": [
                    "Field Definitions"
                ],
                "summary": "Update field definition status",
                "description": "Activate or deactivate a field definition",
                "operationId": "updateStatus_3",
                "parameters": [
                    {
                        "name": "resourceTypeId",
                        "in": "path",
                        "description": "Field definition ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "isActive",
                        "in": "query",
                        "description": "Active status",
                        "required": true,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Field definition not found",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Response"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Status updated successfully",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Response"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/dataloader/configFields/type/{typeName}": {
            "get": {
                "tags": [
                    "Field Definitions"
                ],
                "summary": "Get field definition by type name",
                "description": "Retrieve a field definition by its type name",
                "operationId": "getByTypeName",
                "parameters": [
                    {
                        "name": "typeName",
                        "in": "path",
                        "description": "Type name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Field definition not found",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Response"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Field definition found",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Response"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/dataloader/configFields/provider": {
            "get": {
                "tags": [
                    "Field Definitions"
                ],
                "summary": "Get field definitions by provider",
                "description": "Get all field definitions for a specific provider",
                "operationId": "getByProvider",
                "parameters": [
                    {
                        "name": "provider",
                        "in": "query",
                        "description": "Provider type",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "description": "Provider type enum",
                            "enum": [
                                "CONNECTION",
                                "SOURCE",
                                "TARGET"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Field definitions retrieved successfully",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Response"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid provider",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Response"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/dataloader/configFields/name/{typeName}": {
            "get": {
                "tags": [
                    "Field Definitions"
                ],
                "operationId": "getByTypeNameOld",
                "parameters": [
                    {
                        "name": "typeName",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Response"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ErrorDefinition": {
                "type": "object",
                "properties": {
                    "label": {
                        "type": "string"
                    },
                    "code": {
                        "type": "string"
                    },
                    "level": {
                        "type": "string",
                        "enum": [
                            "REQUEST",
                            "BUSINESS",
                            "SECURITY",
                            "CONFLICT",
                            "RESOURCE"
                        ]
                    },
                    "severity": {
                        "type": "string",
                        "enum": [
                            "FATAL",
                            "NONFATAL"
                        ]
                    },
                    "message": {
                        "type": "string"
                    },
                    "httpStatus": {
                        "type": "string",
                        "enum": [
                            "100 CONTINUE",
                            "101 SWITCHING_PROTOCOLS",
                            "102 PROCESSING",
                            "103 EARLY_HINTS",
                            "103 CHECKPOINT",
                            "200 OK",
                            "201 CREATED",
                            "202 ACCEPTED",
                            "203 NON_AUTHORITATIVE_INFORMATION",
                            "204 NO_CONTENT",
                            "205 RESET_CONTENT",
                            "206 PARTIAL_CONTENT",
                            "207 MULTI_STATUS",
                            "208 ALREADY_REPORTED",
                            "226 IM_USED",
                            "300 MULTIPLE_CHOICES",
                            "301 MOVED_PERMANENTLY",
                            "302 FOUND",
                            "302 MOVED_TEMPORARILY",
                            "303 SEE_OTHER",
                            "304 NOT_MODIFIED",
                            "305 USE_PROXY",
                            "307 TEMPORARY_REDIRECT",
                            "308 PERMANENT_REDIRECT",
                            "400 BAD_REQUEST",
                            "401 UNAUTHORIZED",
                            "402 PAYMENT_REQUIRED",
                            "403 FORBIDDEN",
                            "404 NOT_FOUND",
                            "405 METHOD_NOT_ALLOWED",
                            "406 NOT_ACCEPTABLE",
                            "407 PROXY_AUTHENTICATION_REQUIRED",
                            "408 REQUEST_TIMEOUT",
                            "409 CONFLICT",
                            "410 GONE",
                            "411 LENGTH_REQUIRED",
                            "412 PRECONDITION_FAILED",
                            "413 PAYLOAD_TOO_LARGE",
                            "413 REQUEST_ENTITY_TOO_LARGE",
                            "414 URI_TOO_LONG",
                            "414 REQUEST_URI_TOO_LONG",
                            "415 UNSUPPORTED_MEDIA_TYPE",
                            "416 REQUESTED_RANGE_NOT_SATISFIABLE",
                            "417 EXPECTATION_FAILED",
                            "418 I_AM_A_TEAPOT",
                            "419 INSUFFICIENT_SPACE_ON_RESOURCE",
                            "420 METHOD_FAILURE",
                            "421 DESTINATION_LOCKED",
                            "422 UNPROCESSABLE_ENTITY",
                            "423 LOCKED",
                            "424 FAILED_DEPENDENCY",
                            "425 TOO_EARLY",
                            "426 UPGRADE_REQUIRED",
                            "428 PRECONDITION_REQUIRED",
                            "429 TOO_MANY_REQUESTS",
                            "431 REQUEST_HEADER_FIELDS_TOO_LARGE",
                            "451 UNAVAILABLE_FOR_LEGAL_REASONS",
                            "500 INTERNAL_SERVER_ERROR",
                            "501 NOT_IMPLEMENTED",
                            "502 BAD_GATEWAY",
                            "503 SERVICE_UNAVAILABLE",
                            "504 GATEWAY_TIMEOUT",
                            "505 HTTP_VERSION_NOT_SUPPORTED",
                            "506 VARIANT_ALSO_NEGOTIATES",
                            "507 INSUFFICIENT_STORAGE",
                            "508 LOOP_DETECTED",
                            "509 BANDWIDTH_LIMIT_EXCEEDED",
                            "510 NOT_EXTENDED",
                            "511 NETWORK_AUTHENTICATION_REQUIRED"
                        ]
                    },
                    "sourceApplication": {
                        "type": "string"
                    }
                }
            },
            "Response": {
                "type": "object",
                "properties": {
                    "code": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "type": "object"
                    },
                    "causedBy": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ErrorDefinition"
                        }
                    }
                }
            },
...
            "ConfigField": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "label": {
                        "type": "string"
                    },
                    "required": {
                        "type": "boolean"
                    },
                    "defaultValue": {
                        "type": "object"
                    },
                    "type": {
                        "type": "string"
                    },
                    "placeholder": {
                        "type": "string"
                    },
                    "helpText": {
                        "type": "string"
                    },
                    "order": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "options": {
                        "type": "string"
                    },
                    "min": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "max": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "pattern": {
                        "type": "string"
                    }
                }
            },
            "FieldDefinition": {
                "type": "object",
                "description": "Field definition entity",
                "properties": {
                    "id": {
                        "type": "string",
                        "description": "Unique identifier",
                        "example": "Xkdsancksancsak-ds"
                    },
                    "typeName": {
                        "type": "string",
                        "description": "Type name",
                        "example": "SQL Server"
                    },
                    "provider": {
                        "type": "string",
                        "description": "Provider type",
                        "enum": [
                            "CONNECTION",
                            "SOURCE",
                            "TARGET"
                        ]
                    },
                    "description": {
                        "type": "string",
                        "description": "Description"
                    },
                    "configFields": {
                        "type": "array",
                        "description": "Configuration fields",
                        "items": {
                            "$ref": "#/components/schemas/ConfigField"
                        }
                    },
                    "displayOrder": {
                        "type": "integer",
                        "format": "int64",
                        "description": "Display order"
                    },
                    "isActive": {
                        "type": "boolean",
                        "description": "Active status"
                    },
                    "createdBy": {
                        "type": "string",
                        "description": "Created by"
                    },
                    "createdAt": {
                        "type": "integer",
                        "format": "int64",
                        "description": "Created timestamp"
                    },
                    "updatedBy": {
                        "type": "string",
                        "description": "Updated by"
                    },
                    "updatedAt": {
                        "type": "integer",
                        "format": "int64",
                        "description": "Updated timestamp"
                    },
                    "deletedAt": {
                        "type": "integer",
                        "format": "int64",
                        "description": "Deleted timestamp"
                    }
                }
            },
...
        }
    }
}